Aggregate migration `20241212131910_add_oidc_user_tokens`
